+++ BoxedScene.C
@@ -22,15 +22,15 @@ addObject(SceneObject* s)
    boxes.push_back(new AABB(*s));
 }
 
-double BoxedScene::
+float BoxedScene::
 getLineOfSight(Light& source, const SceneObject& dest, Vector& p)
 {
-   double los = 1.0;
+   float los = 1.0;
 
    /* This is the maximum distance an obscuring object can be at, measured from 
     * 'source'. We subtract a small epsilon value here. The object might fall 
     * inside the MAX_DIST if we don't do this. */
-   const double MAX_DIST = (p - source.pos).getMagnitude() - 0.001f; 
+   const float MAX_DIST = (p - source.pos).getMagnitude() - 0.001f; 
 
    /* A ray from the 'source' vector to the 'dest' vector. */
    Vector& pos = source.pos;
@@ -45,7 +45,7 @@ getLineOfSight(Light& source, const SceneObject& dest, Vector& p)
 
       if (box->intersect(r) && s.intersect(r))
       {
-         double distance = (r.intersection - r.pos).getMagnitude();
+         float distance = (r.intersection - r.pos).getMagnitude();
 
          if (distance < MAX_DIST)
          {
@@ -74,7 +74,7 @@ bool BoxedScene::
 testIntersection(Ray& r)
 {
    bool found = false;
-   double bestDistance = FLT_MAX;
+   float bestDistance = FLT_MAX;
 
    vector<AABB*>::iterator i = boxes.begin();
    while (i != boxes.end())
@@ -86,7 +86,7 @@ testIntersection(Ray& r)
       {
          if (box.object.intersect(temp) == true)
          {
-            double distance = (temp.intersection - temp.pos).getMagnitude();
+            float distance = (temp.intersection - temp.pos).getMagnitude();
 
             if (distance < bestDistance)
             {
